# src/main/resources/swagger.yaml
swagger: '2.0'

info:
  title: User API
  description: User data
  version: "1.0.0"

host: users.dialogue.co

schemes:
  - https

basePath: /v1

produces:
  - application/json

paths:

  /patients:
    get:
      summary: Get paginated patient profiles
      responses:
        200:
          description: |
            Get a page of 10 patient profiles with the link for next page.
          schema:
            type: object
            required:
              - data
              - links
            properties:
              data:
                type: array
                items:
                  $ref: '#/definitions/patient'
              links:
                $ref: '#/definitions/pagination'

    post:
      summary: Creates a patient profile
      responses:
        201:
          description: A patient profile.
          schema:
            $ref: '#/definitions/patient'
        400:
          description: Required field missing or bad format.
          schema:
            $ref: '#/definitions/http-error'
        409:
          description: Email duplicates
          schema:
            $ref: '#/definitions/http-error'
      parameters:
        - name: body
          in: body
          schema:
            type: object
            required:
              - data
            properties:
              data:
                type: object
                required:
                  - email
                  - first_name
                  - last_name
                  - birthdate
                  - sex
                properties:
                  email:
                    type: string
                  first_name:
                    type: string
                  last_name:
                    type: string
                  birthdate:
                    type: string
                    format: date
                  sex:
                    type: string


  /patients/{id}:
    get:
      summary: Get a patient profile
      parameters:
        - name: id
          in: path
          description: the patient id
          required: true
          type: number
      responses:
        200:
          description: patient profile
          schema:
            $ref: '#/definitions/patient'
        404:
          description: not found
          schema:
            $ref: '#/definitions/http-error'

definitions:

  pagination:
    type: object
    required:
      - self
      - next
    properties:
      self:
        type: string
        format: url
        description: Canonical URL of the current page
      next:
        type: string
        format: url
        description: URL of next page

  patient:
    type: object
    required:
      - attributes
      - type
    properties:
      links:
        type: object
        required:
          - self
        properties:
          self:
            type: string
            format: url
            description: Canonical URL of patient profile
      type:
        type: string
        description: "patient"
        x-example: "patient"
      attributes:
        type: object
        required:
          - id
          - email
          - first_name
          - last_name
          - birthdate
          - sex
        properties:
          id:
            type: integer
          email:
            type: string
          first_name:
            type: string
          last_name:
            type: string
          birthdate:
            type: string
            format: date
          sex:
            type: string

  http-error:
    type: object
    properties:
      error:
        type: object
        properties:
          id:
            type: string
            description: Request unique identifier.
          status:
            type: string
            description: HTTP status code.
          title:
            type: string
            description: HTTP status.
          detail:
            type: string
            description: Human readable detail.
          code:
            type: string
            description: An error code unique to the error case.